Abstracted logic from ui to manager
